refactor(web): consolidate Repository Features into Repository Setup
Fold the tri-state Wiki/Issues/Projects/Pull-requests controls into the
Repository Setup section card as a labeled sub-block and drop the standalone
Repository Features section, so all repo-shape config lives in one place. The
four repo_feature_* fields now feed the Repository Setup status badge; the
repo_features wire mapping and schema are unchanged.
